feat(packages): add example monorepo package
- migrates original root package to pnt-functional
- replaces root package with simplified package
