+++ packages/pnt-functional/pyproject.toml
@@ -2,28 +2,33 @@
 authors = [
   { name = "Your Name", email = "your.email@example.com" },
 ]
-dependencies = []
-description = "Add your description here"
-name = "python-nix-template"
+classifiers = [
+  "Development Status :: 3 - Alpha",
+  "License :: OSI Approved :: MIT License",
+  "Operating System :: OS Independent",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+]
+dependencies = [
+  "beartype>=0.19.0,<0.20.0",
+  "expression>=5.5.0",
+]
+description = "A Python package template using Nix and uv"
+license = { text = "Apache-2.0" }
+name = "pnt-functional"
 readme = "README.md"
 requires-python = ">=3.11"
 version = "0.1.0"
 
 [project.scripts]
-python-nix-template = "python_nix_template:main"
+pnt-functional = "pnt_functional.main:greet"
 
 [build-system]
 build-backend = "hatchling.build"
 requires      = ["hatchling"]
 
-# [tool.uv.sources]
-# pnt-functional = { workspace = true }
-
-# [tool.uv.workspace]
-# members = ["packages/*"]
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/python_nix_template"]
+[project.optional-dependencies]
 
 [dependency-groups]
 dev = [
@@ -57,10 +62,12 @@ platforms = ["linux-64", "osx-arm64"]
 preview   = ["pixi-build"]
 
 [tool.pixi.dependencies]
-python = ">=3.11"
+beartype   = ">=0.19.0"
+expression = ">=5.5.0"
+python     = ">=3.11"
 
 [tool.pixi.pypi-dependencies]
-python-nix-template = { path = ".", editable = true }
+pnt-functional = { path = ".", editable = true }
 
 [tool.pixi.environments]
 default     = { solve-group = "default" }
@@ -70,11 +77,26 @@ lint        = { features = ["lint"], solve-group = "default" }
 test        = { features = ["test"], solve-group = "default" }
 types       = { features = ["types"], solve-group = "default" }
 
+[tool.pixi.feature.test.tasks]
+test = "pytest"
+
+[tool.pixi.feature.lint.tasks]
+lint       = "ruff check --fix src/"
+lint-check = "ruff check src/"
+
+[tool.pixi.feature.types.tasks]
+types = "pyright src/"
+
 [tool.pixi.package]
-name    = "python-nix-template"
+name    = "pnt-functional"
 version = "0.1.0"
 
 [tool.pixi.package.run-dependencies]
+beartype   = ">=0.19.0"
+expression = ">=5.5.0"
+
+[tool.pixi.package.host-dependencies]
+hatchling = "*"
 
 [tool.pixi.package.build]
 backend = { name = "pixi-build-python", version = "0.1.*" }
@@ -83,25 +105,21 @@ channels = [
   "https://prefix.dev/pixi-build-backends",
 ]
 
-[tool.pixi.package.host-dependencies]
-hatchling = "*"
+[tool.hatch.build.targets.wheel]
+packages = ["src/pnt_functional"]
 
 [tool.pytest.ini_options]
-# Add source locations to pytest-cov e.g.:
-# --cov=packages/pnt-functional/src/pnt_functional/
 addopts = """
 -rA
 --xdoctest
 --disable-warnings
---cov=src/python_nix_template/
+--cov=src/pnt_functional/
 --cov-report=term-missing
 --durations=0
 """
 log_level = "INFO"
 testpaths = [
-  # Add test locations e.g.:
-  # "packages/pnt-functional/src/pnt_functional/tests",
-  "src/python_nix_template/tests",
+  "src/pnt_functional/tests",
 ]
 
 [tool.ruff]
