+++ packages/pnt-functional/src/pnt_functional/__init__.py
@@ -1,0 +1,9 @@
+"""A Python package template using Nix and uv."""
+
+__version__ = "0.1.0"
+
+from pnt_functional.main import greet
+
+__all__ = [
+    "greet",
+]
