+++ src/python_nix_template/__init__.py
@@ -1,9 +1,2 @@
-"""A Python package template using Nix and uv."""
-
-__version__ = "0.1.0"
-
-from python_nix_template.main import greet
-
-__all__ = [
-    "greet",
-]
+def main() -> None:
+    print("Hello from python-nix-template!")
