+++ src/python_nix_template/tests/__init__.py
@@ -1,0 +1,1 @@
+# This file is intentionally left empty to mark the directory as a Python package
